# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FerreMax CRM/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update sale status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # ─── QUOTES ──────────────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PURCHASES ───────────────────────────────────────────────────────────────
...
  /whatsapp/conversations:
    get:
      operationId: listConversations
      tags: [whatsapp]
      summary: List WhatsApp conversations
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
  /whatsapp/conversations/{id}/messages:
    get:
      operationId: getConversationMessages
      tags: [whatsapp]
      summary: Get conversation messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [whatsapp]
      summary: Send message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── AI ──────────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── REPORTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SaleItem:
      type: object
      required: [productId, productName, quantity, unitPrice, total]
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: ["number", "null"]
        total:
          type: number
    Sale:
      type: object
      required: [id, total, status, paymentMethod, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: ["number", "null"]
        tax:
          type: ["number", "null"]
        total:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [items, paymentMethod]
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, unitPrice]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
              discount:
                type: number
        discount:
          type: number
        tax:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
    SaleUpdate:
      type: object
      properties:
        status:
          type: string
        paymentMethod:
          type: string
        notes:
          type: string

    # ── QUOTE ─────────────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      required: [id, name, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SupplierInput:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        contactName:
          type: string
        notes:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        contactName:
          type: string
        notes:
          type: string

    # ── PURCHASE ──────────────────────────────────────────────────────────────
...
    Conversation:
      type: object
      required: [id, phone, name, lastMessage, lastMessageAt, unreadCount]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        aiEnabled:
          type: boolean
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        label:
          type: ["string", "null"]
    Message:
      type: object
      required: [id, conversationId, content, fromMe, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        content:
          type: string
        fromMe:
          type: boolean
        status:
          type: ["string", "null"]
        messageType:
          type: string
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        messageType:
          type: string

    # ── AI ────────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, role, password]
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean

    # ── REPORTS ───────────────────────────────────────────────────────────────
...